Move utils and relevant tests to own lib package
