+++ .github/workflows/pythonpackage.yaml
@@ -1,0 +1,27 @@
+name: Tests of renewablepotentialslib
+
+on: [push, pull_request]
+
+jobs:
+    test:
+        name: Run tests
+        runs-on: ubuntu-latest
+        defaults:
+            run:
+                shell: bash -l {0}
+        steps:
+            - uses: actions/checkout@v2
+            - name: Setup conda environment
+              uses: conda-incubator/setup-miniconda@v2
+              with:
+                auto-update-conda: true
+                python-version: 3.6
+                add-pip-as-python-dependency: true
+            - name: Install renewablepotentialslib and test requirements.
+              run: |
+                cd ./lib
+                conda env create -f requirements-test.yaml
+                conda activate renewablepotentialslib-test
+            - name: Test with pytest
+              run: |
+                pytest ./lib
