+++ lib/pytest.ini
@@ -1,0 +1,1 @@
+[pytest]
