+++ lib/renewablepotentialslib/__init__.py
@@ -1,0 +1,10 @@
+"""Top-level package of renewablepotentialslib."""
+
+__version__ = "0.1.0" # Additionally defined in setup.py.
+
+# from https://epsg.io/3035
+EPSG_3035 = "EPSG:3035"
+EPSG_3035_PROJ4 = "+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000 +ellps=GRS80 +units=m +no_defs "
+# from https://epsg.io/4326
+WGS84 = "EPSG:4326"
+WGS84_PROJ4 = "+proj=longlat +datum=WGS84 +no_defs "
