+++ lib/renewablepotentialslib/conversion.py
@@ -17,8 +17,7 @@ import shapely.geometry
 import shapely.ops
 import pyproj
 
-# from https://epsg.io/3035
-EPSG_3035_PROJ4 = "+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000 +ellps=GRS80 +units=m +no_defs "
+from renewablepotentialslib import EPSG_3035
 
 
 def watt_to_watthours(watt, duration):
