+++ lib/requirements-test.yaml
@@ -1,0 +1,16 @@
+name:
+    renewablepotentialslib-test
+channels:
+    - conda-forge
+dependencies:
+    - python=3.6
+    - pip=19.1.1
+    - pytest=3.2.3
+    - pycountry=20.7.3
+    - gdal=2.4
+    - rasterio=1.0.25
+    - geopandas=0.8.1
+    - numpy=1.14.6
+    - pandas=0.23.2
+    - pip:
+        -e .
