+++ scripts/capacityfactors/id_map.py
@@ -12,10 +12,7 @@ DTYPE = np.uint16
 NO_DATA_VALUE = 64001
 INDEX_EPSILON = 10e-3
 
-EPSG_3035_PROJ4 = "+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000 +ellps=GRS80 +units=m +no_defs "
-EPSG_3035 = "EPSG:3035"
-WGS84_PROJ4 = "+proj=longlat +datum=WGS84 +no_defs "
-WGS84 = "EPSG:4326"
+from renewablepotentialslib import EPSG_3035, EPSG_3035_PROJ4, WGS84, WGS84_PROJ4
 
 
 def id_map(path_to_timeseries, path_to_map, resolution_km):
