+++ scripts/swiss_building_footprints.py
@@ -5,8 +5,8 @@ from rasterstats import zonal_stats
 import pandas as pd
 import geopandas as gpd
 
-from technical_eligibility import Eligibility
-from conversion import area_in_squaremeters
+from renewablepotentialslib.eligibility import Eligibility
+from renewablepotentialslib.conversion import area_in_squaremeters
 
 
 def swiss_building_footprint(path_to_building_footprint, path_to_eligibility,
