+++ scripts/technical_eligibility.py
@@ -2,85 +2,13 @@
 
 In here, we only exclude areas based on technical restrictions.
 """
-from enum import IntEnum
 
 import numpy as np
 import rasterio
 
 DATATYPE = np.uint8
 
-
-class Eligibility(IntEnum):
-    """Categories defining land eligibility for renewable power."""
-    NOT_ELIGIBLE = 0
-    ROOFTOP_PV = 250
-    ONSHORE_WIND_AND_PV = 180
-    ONSHORE_WIND = 110
-    OFFSHORE_WIND = 40
-
-    @property
-    def area_column_name(self):
-        return "eligibility_{}_km2".format(self.name.lower())
-
-    @staticmethod
-    def onshore():
-        """Returns all onshore eligibilities."""
-        return [
-            Eligibility.NOT_ELIGIBLE,
-            Eligibility.ROOFTOP_PV,
-            Eligibility.ONSHORE_WIND_AND_PV,
-            Eligibility.ONSHORE_WIND,
-        ]
-
-    @staticmethod
-    def offshore():
-        """Returns all offshore eligibilities."""
-        return [
-            Eligibility.OFFSHORE_WIND
-        ]
-
-
-class GlobCover(IntEnum):
-    """Original categories taken from GlobCover 2009 land cover."""
-    POST_FLOODING = 11
-    RAINFED_CROPLANDS = 14
-    MOSAIC_CROPLAND = 20
-    MOSAIC_VEGETATION = 30
-    CLOSED_TO_OPEN_BROADLEAVED_FOREST = 40
-    CLOSED_BROADLEAVED_FOREST = 50
-    OPEN_BROADLEAVED_FOREST = 60
-    CLOSED_NEEDLELEAVED_FOREST = 70
-    OPEN_NEEDLELEAVED_FOREST = 90
-    CLOSED_TO_OPEN_MIXED_FOREST = 100
-    MOSAIC_FOREST = 110
-    MOSAIC_GRASSLAND = 120
-    CLOSED_TO_OPEN_SHRUBLAND = 130
-    CLOSED_TO_OPEN_HERBS = 140
-    SPARSE_VEGETATION = 150
-    CLOSED_TO_OPEN_REGULARLY_FLOODED_FOREST = 160 # doesn't exist in Europe
-    CLOSED_REGULARLY_FLOODED_FOREST = 170 # doesn't exist in Europe
-    CLOSED_TO_OPEN_REGULARLY_FLOODED_GRASSLAND = 180 # roughly 2.3% of land in Europe
-    ARTIFICAL_SURFACES_AND_URBAN_AREAS = 190
-    BARE_AREAS = 200
-    WATER_BODIES = 210
-    PERMANENT_SNOW = 220
-    NO_DATA = 230
-
-
-FARM = [GlobCover.POST_FLOODING, GlobCover.RAINFED_CROPLANDS,
-        GlobCover.MOSAIC_CROPLAND, GlobCover.MOSAIC_VEGETATION]
-FOREST = [GlobCover.CLOSED_TO_OPEN_BROADLEAVED_FOREST, GlobCover.CLOSED_BROADLEAVED_FOREST,
-          GlobCover.OPEN_BROADLEAVED_FOREST, GlobCover.CLOSED_NEEDLELEAVED_FOREST,
-          GlobCover.OPEN_NEEDLELEAVED_FOREST, GlobCover.CLOSED_TO_OPEN_MIXED_FOREST,
-          GlobCover.MOSAIC_FOREST, GlobCover.CLOSED_TO_OPEN_REGULARLY_FLOODED_FOREST,
-          GlobCover.CLOSED_REGULARLY_FLOODED_FOREST]
-VEGETATION = [GlobCover.MOSAIC_GRASSLAND, GlobCover.CLOSED_TO_OPEN_SHRUBLAND,
-              GlobCover.CLOSED_TO_OPEN_HERBS, GlobCover.SPARSE_VEGETATION,
-              GlobCover.CLOSED_TO_OPEN_REGULARLY_FLOODED_GRASSLAND]
-BARE = [GlobCover.BARE_AREAS]
-OTHER = VEGETATION + BARE
-URBAN = [GlobCover.ARTIFICAL_SURFACES_AND_URBAN_AREAS]
-WATER = [GlobCover.WATER_BODIES]
+from renewablepotentialslib.eligibility import Eligibility, FARM, FOREST, OTHER, WATER
 
 
 def determine_eligibility(path_to_land_cover, path_to_slope, path_to_bathymetry,
